LayoutTests/http/tests/media/video-accept-encoding-expected.txt

Test that video requests contain no "Accept-Encoding" header.

EVENT(loadedmetadata)
loaded metadata of media file OK
END OF TEST
